+++ script/cleanup-all.ts
